// om-ddm-billing-confirmation-service/src/proto/billing_confirmation.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.6.1
//   protoc               v5.29.2
// source: billing_confirmation.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "billing_confirmation";

export interface GetBillingRequest {
  invoiceId: string;
}

export interface GetBillingResponse {
  invoiceId: string;
  amount: number;
  isConfirmed: boolean;
}

export interface CreateBillingRequest {
  invoiceId: string;
  amount: number;
}

export interface CreateBillingResponse {
  success: boolean;
  message: string;
}

export const BILLING_CONFIRMATION_PACKAGE_NAME = "billing_confirmation";

export interface BillingConfirmationServiceClient {
  getBilling(request: GetBillingRequest): Observable<GetBillingResponse>;

  createBilling(request: CreateBillingRequest): Observable<CreateBillingResponse>;
}

export interface BillingConfirmationServiceController {
  getBilling(
    request: GetBillingRequest,
  ): Promise<GetBillingResponse> | Observable<GetBillingResponse> | GetBillingResponse;

  createBilling(
    request: CreateBillingRequest,
  ): Promise<CreateBillingResponse> | Observable<CreateBillingResponse> | CreateBillingResponse;
}

export function BillingConfirmationServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["getBilling", "createBilling"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("BillingConfirmationService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("BillingConfirmationService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const BILLING_CONFIRMATION_SERVICE_NAME = "BillingConfirmationService";